move the null epoch definition to epoch module and add check for it when validing genesis check whether the completed epoch number of undelgation is negative when validing genesis file
